Remove support for using a random salt
Since using a random salt only works if you're using a hash function
such as SipHash and we have no working implementation of SipHash,
providing a random salt mechanism provides users with a false sense of
security, which is worse than no security at all. Now we instead
document the limitation.

We will most likely re-add support for a random salt once we have a
working version of SipHash. When we re-add it we should use a
newtype (e.g. Salted) for users to use to opt in to the behavior. Using
a flag such as -ffixed-salt doesn't work well, as users often receive
hashable precompiled as part of the Haskell Platform, which doesn't
set this flag to false.

An option to using a Salted newtype is to use a SafeByteString newtype
that combines using a random salt with using a stronger hash
function (i.e. SipHash).