Change HashArgs to be data-family
Now, as HashArgs doesn't carry index evidence, pattern matching on
`HashArgs` doesn't bring any evidence into the scope.
We need to know the index first to be able to pattern match.
This is small cosmetic detail, as in practice GHC is smart enough;
but maybe GHC will be tiny-bit faster generating Generic instances.